+++ filter_xargsbasename.go
@@ -39,19 +39,34 @@
 
 package scriptish
 
-import "os"
+import (
+	"path/filepath"
+	"strings"
+)
 
-// Exit terminates the Golang app with the given status code.
-//
-// It does *NOT* flush the pipe's Stdout or Stderr to your Golang's
-// os.Stdout / os.Stderr first.
-func Exit(statusCode int) Command {
+// XargsBasename treats every line in the pipe as a filepath.
+// It removes any parent elements from the line.
+func XargsBasename() Command {
 	// build our Scriptish command
 	return func(p *Pipe) (int, error) {
-		// all done
-		os.Exit(statusCode)
+		// process each filepath in the pipeline
+		for line := range p.Stdin.ReadLines() {
+			var basename string
 
-		// this is unreachable, but added to keep the compiler happy
-		return statusCode, nil
+			if len(strings.TrimSpace(line)) > 0 {
+				// let's get our basename
+				basename = filepath.Base(line)
+			} else {
+				// special case - preserve blank lines
+				basename = ""
+			}
+
+			// send what we've got
+			p.Stdout.WriteString(basename)
+			p.Stdout.WriteRune('\n')
+		}
+
+		// all done
+		return StatusOkay, nil
 	}
 }